requirements.txt
streamlit>=1.30.0
google-generativeai>=0.5.0
numpy>=1.24.0